requirements.txt
streamlit
PyMuPDF
google-genai
pytest
pytesseract
Pillow
